Pull back hard fork
Due to us hitting block 62500 a lot sooner than I anticipated and not
being able to get new clients in place fast enough I've had to go with
the majority of the users and checkpoint the fork mined with the older
clients.

As such I've moved these changes to come into force at block 64000
instead.

Apologies for any inconvenience if you updated ahead of this and ended
up on the now invalidated fork.