GameCoin rollback
Rolling back to block 63,999 as the last block that was accepted
globally before the previous hard fork came into affect.

A new block 64,000 has already been mined and has been added as a
checkpoint but the block will not be available on the network until the
pools are up to date and the new client is compiled and available to
all.

I have decided to reverse my decision to reduce the block times to 1
minute intervals and reduce the coin subsidy to 400 GME. Instead we will
continue with 2.5 minute blocks but reduce the retarget interval from 48
blocks to 12.

From the time block 64,000 is actually released on the network I
anticipate the difficulty will climb to ~0.25 by block 64068, at which
point it should increase/decrease by no more than 110% at each retarget.