Hard Fork at block 69000
Bring the rolling average down from 48 retarget intervals (576 blocks)
to 4 retarget intervals (48 blocks) as looking a theoretical 24 hours
into the block history results in a very delayed response to changes in
the network speed.

Also switch to the same 48 block timeframe for calculating the network
speed as 576 blocks results in the same delayed response in reporting
the network speed.